Rename config to entry and store to config
